Dispatch dom related methods to helpers.dom
