+++ src/helpers/index.js
@@ -5,3 +5,4 @@ module.exports.easing = require('./helpers.easing');
 module.exports.canvas = require('./helpers.canvas');
 module.exports.options = require('./helpers.options');
 module.exports.math = require('./helpers.math');
+module.exports.dom = require('./helpers.dom');
